[frontend] investigation 2
