+++ opencti-platform/opencti-front/src/private/components/workspaces/investigations/Investigation.jsx
@@ -5,8 +5,7 @@ import { graphql, createFragmentContainer } from 'react-relay';
 import withStyles from '@mui/styles/withStyles';
 import { QueryRenderer } from '../../../../relay/environment';
 import inject18n from '../../../../components/i18n';
-import InvestigationKnowledgeGraph, { investigationGraphQuery } from './InvestigationGraph';
-import InvestigationGraph from './_InvestigationGraph';
+import InvestigationGraph, { investigationGraphQuery } from './InvestigationGraph';
 import Loader from '../../../../components/Loader';
 import ErrorNotFound from '../../../../components/ErrorNotFound';
 
@@ -31,7 +30,6 @@ class InvestigationComponent extends Component {
             if (props) {
               if (props.workspace) {
                 return (
-                  // <InvestigationKnowledgeGraph workspace={props.workspace} />
                   <InvestigationGraph data={props.workspace} graphData={props.workspace} />
                 );
               }
